Fixes trac:3044: Improve the collective algorithm selection
---svn-pre-commit-ignore-below---
r26080
Help he selection of the right algorithm for large data (> 2Gb). Thanks
to Fujitsu for the patch.

This commit was SVN r26126.

The following SVN revision numbers were found above:
  r26080 --> open-mpi/ompi@8791ade2935053801a1e4da9b1c2b52f4e76799c

The following Trac tickets were found above:
  Ticket 3044 --> https://svn.open-mpi.org/trac/ompi/ticket/3044

+++ ompi/mca/coll/tuned/coll_tuned_dynamic_rules.c
@@ -9,6 +9,7 @@
  *                         University of Stuttgart.  All rights reserved.
  * Copyright (c) 2004-2005 The Regents of the University of California.
  *                         All rights reserved.
+ * Copyright (c) 2011-2012 FUJITSU LIMITED.  All rights reserved.
  * $COPYRIGHT$
  * 
  * Additional copyrights may follow
@@ -346,7 +347,7 @@ ompi_coll_com_rule_t* ompi_coll_tuned_get_com_rule_ptr (ompi_coll_alg_rule_t* ru
  *
  */
 
-int ompi_coll_tuned_get_target_method_params (ompi_coll_com_rule_t* base_com_rule, int mpi_msgsize, int *result_topo_faninout, 
+int ompi_coll_tuned_get_target_method_params (ompi_coll_com_rule_t* base_com_rule, size_t mpi_msgsize, int *result_topo_faninout, 
                                               int* result_segsize, int* max_requests)
 {
     ompi_coll_msg_rule_t*  msg_p = (ompi_coll_msg_rule_t*) NULL;
